Fix booleans *again*, for real this time
Previous fix was operating under a false assumption; after doing a
little debugging, the Buildkite Agent expresses these values as the
strings "true" or "false". Our test should take that into account.

This is implied by json-schema.org's guidlines for boolean values but
it is not made clear in the Buildkite plugin documentation.