Booleans, Pt. 3: Bool Harder
It looks like the values of these options are passed through literally
as environment variables with no casting whatsoever done. So we need
to accommodate for "true" and "1" and case insensitivity when figuring
out if shopts should be set. There's some test cases to flex the
negative scenarios, as well as the positive.