Refactor testing for valid shopts
Instead of testing against a version lookup, we can just try setting
the shopt directly and handling the failure ourselves. It's a little
more brute-force but it's also a little bit less dependent on internal
Bash variables.

The nature of shopts means we can't directly test them without
overriding so much code that we're essentially using a function to
exercise a test (not the other way around). But there's ample comments
describing the important details, like "no major platform ships bash
3 except macOS".